Rename files for array expression conversions in order to avoid naming conflicts in TAB-completion of the corresponding functions
