+++ sympy/matrices/expressions/matexpr.py
@@ -442,8 +442,8 @@ class MatrixExpr(Expr):
         >>> MatrixExpr.from_index_summation(expr)
         A*B.T*A.T
         """
-        from sympy.tensor.array.expressions.conv_indexed_to_array import convert_indexed_to_array
-        from sympy.tensor.array.expressions.conv_array_to_matrix import convert_array_to_matrix
+        from sympy.tensor.array.expressions.from_indexed_to_array import convert_indexed_to_array
+        from sympy.tensor.array.expressions.from_array_to_matrix import convert_array_to_matrix
         first_indices = []
         if first_index is not None:
             first_indices.append(first_index)
@@ -522,9 +522,9 @@ def _matrix_derivative(expr, x, old_algorithm=False):
     if old_algorithm:
         return _matrix_derivative_old_algorithm(expr, x)
 
-    from sympy.tensor.array.expressions.conv_matrix_to_array import convert_matrix_to_array
+    from sympy.tensor.array.expressions.from_matrix_to_array import convert_matrix_to_array
     from sympy.tensor.array.expressions.arrayexpr_derivatives import array_derive
-    from sympy.tensor.array.expressions.conv_array_to_matrix import convert_array_to_matrix
+    from sympy.tensor.array.expressions.from_array_to_matrix import convert_array_to_matrix
 
     array_expr = convert_matrix_to_array(expr)
     diff_array_expr = array_derive(array_expr, x)
@@ -538,7 +538,7 @@ def _matrix_derivative_old_algorithm(expr, x):
 
     parts = [i.build() for i in lines]
 
-    from sympy.tensor.array.expressions.conv_array_to_matrix import convert_array_to_matrix
+    from sympy.tensor.array.expressions.from_array_to_matrix import convert_array_to_matrix
 
     parts = [[convert_array_to_matrix(j) for j in i] for i in parts]
