+++ sympy/tensor/array/expressions/from_array_to_indexed.py
