+++ sympy/tensor/array/expressions/from_matrix_to_array.py
